fix: removes cast function definition (#253)
Given that the `Expression` node has an explicit `Cast` variant that was
added in #88, the `cast` function definition (added in #152) seems redundant.

BREAKING CHANGE: Existing plans that are modeling `cast` with the `cast`
function (as opposed to the `cast` expression) will no longer be valid. All
producers/consumers should use the `cast` expression type.