feat(set): add `index_in` set membership operation
Adds an `index_in` function to determine whether a value is a
member of a given set of values and the index of that value if it
exists.

I put this in a new yaml file to start with, but I think we could
_probably_ stick it in `comparisons`?  Open to thoughts on where this
belong.  If we are planning to add other `set` operations (union,
difference, intersection, etc) then we may be better served by having a
separate file.