Add Apply relational operator support to substrait
This PR adds a new relation type for APPLY operation. Apply is similar
to JOIN operator since it joins two table sources. However the table
source on the right depends on computation of the table source on right.
Since the right source cannot be computed without right, it is
represented by an Expression and not a relation. In Substrait the right
source will most likely be a SubQuery.