feat: clarify map behavior
   * Map keys must be unique.
   * Map keys must not be NULL.
   * The map key type may be nullable.

This is based on the current restrictions found in the wild.  For example, DuckDB specifically calls out that keys must be unique in its implementation and Spark specifically calls out that keys may not be NULL.  This makes sense as map keys often need to be hashable for performance reasons.

If for some reason we need a map type that supports repeated keys it may be better to create an alterate type (or a clarifying parameter/option).