feat: add missing rels to rel message (#582)
The `ReferenceRel`, `WriteRel`, and `DdlRel` were defined in
`algebra.proto` but not part
of `message Rel` which meant they were unusable. This PR adds those
back. It is
inspired by #288 but more targeted in scope. One change from that
original PR which I
also kept was replacing the word `tuple` with `record` in the
documentation for consistency.

This is not to imply that `ReferenceRel`, `WriteRel`, or `DdlRel` are
"complete" or "stable"
in any way. I feel these relations are still quite ill defined. However,
my hope is that by
making them usable we can inspire further change to them.

BREAKING CHANGE: The enum `WriteRel::OutputMode` had an option change
from
`OUTPUT_MODE_MODIFIED_TUPLES` to `OUTPUT_MODE_MODIFIED_RECORDS`
BREAKING CHANGE: The message `AggregateFunction.ReferenceRel` has moved
to `ReferenceRel`.