fix: include precision information in PrecisionTimestamp and PrecisionTimestampTZ literals (#659)
BREAKING CHANGE: changes the message type for Literal PrecisionTimestamp
and PrecisionTimestampTZ

The PrecisionTimestamp and PrecisionTimestampTZ literals were introduced
in #594, and there was some discussion about their contents in
https://github.com/substrait-io/substrait/pull/594#discussion_r1471844566.
In their current form, they only contain a i64 value, which I believe is
meant to be a precision-dependent number of fractional seconds since
epoch. However, the literals don't contain the precision itself, so it's
impossible to interpret a PrecisionTimestamp or PrecisionTimestampTZ
literal without other information. This is in contrast to e.g. varchar,
whose literal does specify the length, or decimal, whose literal
specifies scale and precision. @westonpace curious for your thoughts
since you were part of that original discussion - am I missing something
or is this a bug?

This PR changes the Literal types for PrecisionTimestamp and
PrecisionTimestampTZ to contain a PrecisionTimestamp message instead of
a i64. That message then contains the i64 value as well as the i32 type.
This is a breaking change, but given in their current form these
literals aren't usable (unless I'm missing something), would that be
okay?

Currently I used the same message for both PrecisionTimestamp and
PrecisionTimestampTZ, but I can also make a copy for PTTZ if that'd be
better for forwards-compatibility.