feat: add precision to IntervalDay and new IntervalCompound type (#665)
* Update IntervalDay to support multiple subsecond precisions (0-9).
This is done in a way that should be effectively compatible with old
systems. Old plans should be able to be continue to be consumed treating
the old values as microsecond precision.
* Add new IntervalCompound type which is a combination of IntervalMonth
and IntervalDay

BREAKING CHANGE: The encoding of IntervalDay literals has changed in a
strictly backwards incompatible way. However, the logical meaning across
encoding is maintained using a oneof. Moving a field into a oneof makes
unset/set to zero unclear with older messages but the fields are defined
such that the logical meaning of the two is indistinct. If neither
microseconds nor precision is set, the value can be considered a
precision 6 value. If you aren't using IntervalDay type, you will not
need to make any changes.

BREAKING CHANGE: TypeExpression and Parameterized type protobufs (used
to serialize output derivation) are updated to match the now compound
nature of IntervalDay. If you use protobuf to serialize output
derivation that refer to IntervalDay type, you will need to rework that
logic.

fixes #664

---------

Co-authored-by: Jacques Nadeau <jacques@apache.org>