feat: add decimal argument support to round function
The round function has a number of variants to support different numeric types.  This commit adds support for rounding decimals.

The precision of the resultant decimal type is one greater than
the precision of the input decimal to allow for rounding up to the
next decimal digit.
The scale of the resultant decimal type is the same as the input type since the result of rounding cannot add any further decimal places.

Signed-off-by: Andrew Coleman <andrew_coleman@uk.ibm.com>