docs: add definition and voting rules of `-contrib` repositories (#707)
This PR adds the concept of `-contrib` repositories to the governance
rules of Substrait. To that aim, the following changes are made:

* Mention that there is a distinction between "core" and `-contrib`
repositories in the introduction of the Substrait project. Those
definitions are used in what follows.
* Say that contributors can be given write access to `-contrib`
repositories by the SMC. This allows people to make changes to those
repositories without being promoted to committers.
* Clarify that committers are people who have write access to all
repositories, i.e., "core" and `-contrib` repositories. This makes clear
that having write access to a `-contrib` repository doesn't make someone
a committer.
* Add a voting rule for `-contrib` repositories:
* The SMC needs to vote to create new `-contrib` repos and to give write
access to them.
* Any contributor other than the proposer can vote. This is a pretty low
bar but it ensures that one other human being with some relationship to
the project approves the change.

---------

Signed-off-by: Ingo Müller <ingomueller@google.com>
Co-authored-by: Jacques Nadeau <jacquesnadeau@gmail.com>
Co-authored-by: Weston Pace <weston.pace@gmail.com>