feat: anchor for non-ordinal reference to PlanRel
This adds an "anchor" to `PlanRel` that can be referenced from a
`ReferenceRel`. This anchor and reference relationship provides a
non-ordinal mechanism for identifying and accessing a "subtree" or
sub-graph.

PlanRel. The referenced attribute, `reference_anchor`, is named to
specify it is an anchor--a fixed value that can be referenced. The
default value (0) means it is unset and implies it must be actively set
by a producer.

ReferenceRel. The existing field, `subtree_ordinal`, is deprecated. The
reference attribute, `planrel_reference`, is named to specify it is a
reference to a `PlanRel` message. The default value (0) means it is
unset and implies it must be actively set by a producer.

Note: The default value of `subtree_ordinal` (0) can look valid even
when unset. If `planrel_reference` is unset: (1) during the deprecation
period, `subtree_ordinal` can be used as a fallback; (2) in the
future, the `ReferenceRel` message should be considered invalid.

Issue: #725
PR: 726