docs: fix some usages of the word "dataset" (#760)
This PR fixes several usages of the word "dataset". In one occasion, the
word "transformation" was used when really the *result* of such a
transformation was meant, for which the spec uses the word "dataset". In
two other occasions, the word "dataset" was used when really the
*transformation* that produces such a dataset was meant, for which the
spec uses the words "transformation", "relation", "operation", or
"operator."

Signed-off-by: Ingo Müller <ingomueller@google.com>