docs: fix and simplify logic that generates extension options docs (#761)
This PR fixes the logic in `generate_function_docs.py` that generates
the docs for the options of functions. The previous version mixed up the
option names and the corresponding choices, leading to non-sensical docs
such as reported in #755. This PR collects the option names and
corresponding choices jointly such that no mix-up can happen. I have
looked at the output that is changed by this script but only
superficially so: my impression is that (1) some options now come in a
different order (because the previous order was determined by the order
of keys in a `dict`, which is implementation-defined) and (2) the
remaining changes come from the intention of this PR, i.e., the
constitute fixes of previously non-sensical output. However, since I am
not intrically familiar with all extension functions, I am not 100%
sure.

This resolves #755.

Signed-off-by: Ingo Müller <ingomueller@google.com>