feat: specify build input of hash join operator (#810)
# Background

Substrait declares build is right and probe is left today. Unfortunately
there are other systems (SQL Server family, Oracle, Snowflake, Google
Cloud Spanner, etc) chose opposite side and it is a major headache for
those systems to deal with the mirror images.

# Proposal

Adding a `build_left` field to signify which side is the build of the
hash join. Default (false) implies the current behavior. Left and right
does not change. Spark SQL/Spanner employs this approach and be
flexible.

The change is backward/binary compatible.