feat: canonicalize extension URNs (#859)
Introduces the notion of URNs to uniquely identify extensions.

As part of the migration towards URNs and away from the current URIs,
all references to URIs have a corresponding URN reference introduced in
a neighboring field. The URI-fields have all been marked as deprecated.

All usages of URIs will be dropped in future work after a migration
period.