docs: clarify meaning of LocalFiles (#871)
Closes #869 

This change clarifies the documentation surrounding `LocalFiles`. It
will be clearer that `LocalFiles` can also include references to blob
storage etc. as specified by the URI's protocol.

It also moves the duplication of the `ReadRel` proto in the
documentation to the end of the section to be consistent with other
`*Rel`s documetation.